[#551][Refactor] 멤버랭킹 조회 쿼리 오류 수정

+++ src/main/java/com/runninghi/runninghibackv2/application/dto/memberchallenge/response/ChallengeRankResponse.java
@@ -5,6 +5,6 @@ public record ChallengeRankResponse(
         float record,
         String nickname,
         String profileImageUrl,
-        int rank
+        Long rank
 ) {
 }
